Fix Windows integration tests
This commit fixes issues with Windows integration tests and enables the
tests to run through github. The problem was that the tests using the
log_dir option expected symlinks to the log files to exist, those
symlinks being (main.INFO, main.WARN, etc). This was a problem for
two reasons:

1) On Windows the symlinks would start with main.exe, not just main
2) The creation of symlinks is not guarenteed on any platform, and on
Windows creating a symlink requires Administrator permissions.

This was fixed by changing the tests to search for either the symlinks
(with a fixed filepath), or the log files themselves.

Future work can be done to deal with the lack of symlinks on Windows,
but this commit's purpose is to get tests in a passing state.

(cherry picked from commit 5727d2aa060c7b678d18b57c3e977e63ebfb1c6f)